move Chunk and ChunkData to mars/core
